Update image schema tests
